Log `Error#cause` in `dev`
Previously, the `cause` property of errors was ignored when logging.
This is used by D1 to include context on why an operation failed.
Because we weren't logging this, D1 failures would always show up as
`D1_ERROR`, with no additional context, making debugging difficult.

This change attempts to extract `cause` from the object preview sent
back from the inspector with the rest of exception details.
Unfortunately, the stack trace we get back is an unstructured string,
so we need to parse it into an array of call sites before we can
source-map. The stack trace may be truncated as this is only a
preview. In this case, we just log the truncation. We could send a
request to the inspector to get the full stack in this case if we
wanted. Additionally, the cause is sometimes not returned on the
first request. This is the same bug that causes objects entries not
to be displayed when `console.log`ging on first request too.